Разнёс утилитарный модуль
