+++ src/utils/common.js
@@ -1,0 +1,8 @@
+const getRandomInteger = (a = 1, b = 0) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+export {getRandomInteger};
